Redesign of FetchedModelsController
Notifications are now observed in delegate didSet.
modelContext is now let.
New internal struct for caching the state the fetch was started with which means changing the vars like sort/predicate don't affect the previous fetch updates.
Switched from @Observable back to Combine so all the @ObservationIgnored can be removed